Update mdbook, pulldown-cmark and pulldown-cmark-to-cmark dependencies (#18)
This fixes issue #17, which would cause `cargo build mdbook-graphviz` to
fail. This brings packages up to their current latest versions:

* pulldown-cmark: From 0.7 to 0.8
* pulldown-cmark-to-cmark: From 5.0 to 7.1
* mdbook: From 0.4.x to ^0.4.14

The mdbook update is required because this version added a parameter
named `curly_quotes` to the `new_cmark_parser` function. I elected to
set this parameter to `false`, as there does not appear to be
configuration in the mdbook package for this yet. Presumably, that will
come in the future and it should be copied into this repository.